Add files via upload
I created this power BI dashboard to showcase my skills in using Power BI for Data Visualization. I used an HR data set to Create the visualization as I was able to determine the employees due for promotion and the ones to be laid off. The results gotten from the analysis is displayed on the dashboard as shown.